// frontend/src/types/api.ts
export interface components {
    schemas: {
// ...
        JobOut: {
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /**
             * Client Id
             * Format: uuid
             */
            client_id: string;
            /** Provider Id */
            provider_id?: string | null;
            /** Service Listing Id */
            service_listing_id?: string | null;
            /** Title */
            title: string;
            /** Description */
            description: string;
            /** Status */
            status: string;
            /** Price */
            price: string;
            /** Min Price */
            min_price?: string | null;
            /** Max Price */
            max_price?: string | null;
            /** Latitude */
            latitude?: number | null;
            /** Longitude */
            longitude?: number | null;
            /** Escrow Address */
            escrow_address?: string | null;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /** Updated At */
            updated_at?: string | null;
            /** Image Url */
            image_url?: string | null;
        };
// ...
        ServiceOut: {
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /**
             * Provider Id
             * Format: uuid
             */
            provider_id: string;
            /**
             * Category Id
             * Format: uuid
             */
            category_id: string;
            /** Title */
            title: string;
            /** Description */
            description?: string | null;
            /** Price */
            price: string;
            /** Min Price */
            min_price?: string | null;
            /** Max Price */
            max_price?: string | null;
            /** Is Active */
            is_active: boolean;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /** Updated At */
            updated_at?: string | null;
            /** Latitude */
            latitude: number;
            /** Longitude */
            longitude: number;
            /** Radius Km */
            radius_km: string;
            /** Image Url */
            image_url?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}